IntegerField to FloaoaoaoaoatField

+++ homepage/models.py
@@ -95,12 +95,12 @@ class Design(models.Model):
 
 class Logo(models.Model):
     src = models.TextField(default=DEFAULT_LOGO_BASE64, blank=True, null=True)
-    width = models.IntegerField(default=571)
-    height = models.IntegerField(default=589)
-    left = models.IntegerField(default=0)
-    top = models.IntegerField(default=0)
-    scaleX = models.IntegerField(default=1)
-    scaleY = models.IntegerField(default=1)
+    width = models.FloatField(default=571)
+    height = models.FloatField(default=589)
+    left = models.FloatField(default=0)
+    top = models.FloatField(default=0)
+    scaleX = models.FloatField(default=1)
+    scaleY = models.FloatField(default=1)
 
 class Text(models.Model):
     textvalue = models.CharField(max_length=50)
@@ -108,17 +108,17 @@ class Text(models.Model):
     fill = models.CharField(max_length=50)
     fontStyle = models.CharField(max_length=50)
     fontSize = models.IntegerField(default=100)
-    left = models.IntegerField(default=0)
-    top = models.IntegerField(default=0)
+    left = models.FloatField(default=0)
+    top = models.FloatField(default=0)
     stroke = models.CharField(
         max_length=7,
         default="#fcfcfc",
     )
     strokeWidth = models.IntegerField(default=0)
-    scaleX = models.IntegerField(default=1)
-    scaleY = models.IntegerField(default=1)
-    width = models.IntegerField(default=0)
-    height = models.IntegerField(default=0)
+    scaleX = models.FloatField(default=1)
+    scaleY = models.FloatField(default=1)
+    width = models.FloatField(default=0)
+    height = models.FloatField(default=0)
 
 
 class Profile(models.Model):
